feat: add base64 encode/decode filters
This implements `base64encode` and `base64decode` filters.

This is done mainly for the use case of static stite generators such as gutenberg. For this cases, pretty often you need to implement more dynamic features that rely on current data and not in data available at the moment of the html render. For example, being able to manipulate content in a posts index based on the comparison of the post's date and the current browser's date in the client. However, when giving data processing control to the client, you can end up having difficulties handling multiline content or custom data. In such cases it's possible to use base64 to encode the data in the render stage and have the browser decode the data when needed.