escape_html: stop encoding / as &#x2F; (+ more)
The OWASP XSS cheat sheet is a very badly-written document that hasn’t
been maintained and still contains various errors that have lasted more
than a decade, in some cases despite them being pointed out. Also the
section being quoted here was being misapplied anyway (it’s only for
*text*, not for attribute values, and it therefore escapes *way* more
than is needed). The entire document urgently needs to be completely
rewritten, but they’re not doing it. Hence in part my removal of any
citation of it.

One recently exorcised ancient error is the recommendation to escape
slashes: <https://github.com/OWASP/CheatSheetSeries/issues/515>. That
was *always* spurious, and I want it gone partly under the principle of
least encoding but mostly because I’m fed up with URLs being uglified in
this way.

I’ve also changed the escaping of ' from &#x27; to &apos;, because the
reason for avoiding &apos; is invalid (it was an accidental omission in
an early HTML5 spec, long since reinstated, and all user agents always
supported it).