feat: add `Tera::new_unitialized` (#677)
Exposes a new function which creates an instance of Tera without
registering the default filters, testers, and functions. If this is
merged, the version of `tera` in `Cargo.toml` will have to get
its minor version incremented. No tests were added as no new
code was actually; I simply cut and paste the body from the
`Default` implementation. I need this functionality and thing no-ops
as an overwrite are too hacky. I hope others can benefit from this, too.