Add a "per_render" Context option
In the case you have a particularly large Context that you can't easily
clone, it is useful to have a the ablity to augment the Context on a
particular render. By adding an extra context that can be specified
on each render call the large Context can be shared with other render
calls across multiple threads, thus sidestepping the need to clone and
modify the Context for each thread.

I, somewhat aribrary, choose the `per_render_context` to be a fallback context,
if the value is not found in the `Context` then look in the `per_render_context`.
It might be desirable to check for overlapping keys and error out if they exist.